requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cloudinary==1.28.0
dj3-cloudinary-storage==0.0.6
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
mysqlclient==2.0.3
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
waitress==2.0.0
whitenoise==5.3.0
